requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.3
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
phonenumbers==8.13.31
PyJWT==2.8.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
